+++ mint-inference/src/main/java/mint/inference/gp/tree/NodeSimplifier.java
@@ -56,8 +56,7 @@ public class NodeSimplifier {
 			return (Node<IntegerVariableAssignment>) makeBinary(exp.getArgs(), new MultiplyIntegersOperator());
 		}
 		if (exp.isIntNum()) {
-			IntegerVariableAssignment num = new IntegerVariableAssignment(exp.toString(),
-					Integer.valueOf(exp.toString()));
+			IntegerVariableAssignment num = new IntegerVariableAssignment(exp.toString(), Long.valueOf(exp.toString()));
 			return new IntegerVariableAssignmentTerminal(num, true, false);
 		}
 		if (exp.isConst()) {
