+++ mint-inference/src/main/java/mint/inference/gp/tree/nonterminals/booleans/LTBooleanIntegersOperator.java
@@ -56,7 +56,7 @@ public class LTBooleanIntegersOperator extends BooleanNonTerminal {
 	public BooleanVariableAssignment evaluate() throws InterruptedException {
 		checkInterrupted();
 		BooleanVariableAssignment res = new BooleanVariableAssignment("result",
-				(Integer) children.get(0).evaluate().getValue() < (Integer) children.get(1).evaluate().getValue());
+				(Long) children.get(0).evaluate().getValue() < (Long) children.get(1).evaluate().getValue());
 		vals.add(res.getValue());
 		return res;
 	}
