+++ mint-inference/src/main/java/mint/inference/gp/tree/nonterminals/integers/IntegerNonTerminal.java
@@ -33,7 +33,7 @@ public abstract class IntegerNonTerminal extends NonTerminal<IntegerVariableAssi
 
 	@Override
 	public Terminal<IntegerVariableAssignment> getTermFromVals() {
-		IntegerVariableAssignment ivar = new IntegerVariableAssignment("res", (Integer) vals.iterator().next());
+		IntegerVariableAssignment ivar = new IntegerVariableAssignment("res", (Long) vals.iterator().next());
 		IntegerVariableAssignmentTerminal term = new IntegerVariableAssignmentTerminal(ivar, true, false);
 		return term;
 	}
